refactor(front): Addhexagonal architecture
