+++ guifaffinity-front/src/app/App.test.tsx
@@ -2,10 +2,11 @@ import { render, screen } from "@testing-library/react";
 import { App } from "./App";
 import { server } from "../mocks/server.js";
 import { rest } from "msw";
-import { FetchGifsRepository } from "../repositories/FetchGifsRepository";
+import { FetchGifsRepository } from "../infrastructures/FetchGifsRepository";
+import { GifDTO } from "../infrastructures/GifDTO";
 
 test("displays a gif", async () => {
-  const retrievedGif = {
+  const retrievedGif: GifDTO = {
     id: "1",
     src: "https://media.giphy.com/media/xT77XZrTKOxycjaYvK/giphy.gif",
     likes: 15,
