+++ guifaffinity-front/src/domain/Gif.ts
@@ -1,0 +1,8 @@
+export interface Gif {
+  id: string;
+  src: string;
+  likes: number;
+  date: string;
+  alt: string;
+  tags: string[];
+}
