+++ guifaffinity-front/src/domain/GifsRepository.ts
@@ -1,0 +1,5 @@
+import { Gif } from "./Gif";
+
+export interface GifsRepository {
+  getGifs: (tag: string) => Promise<Gif[]>;
+}
