+++ guifaffinity-front/src/infrastructures/FetchGifsRepository.ts
@@ -1,0 +1,11 @@
+import { Gif } from "../domain/Gif";
+import { GifsRepository } from "../domain/GifsRepository";
+import { GifDTO } from "./GifDTO";
+
+export class FetchGifsRepository implements GifsRepository {
+  async getGifs(tag: string): Promise<Gif[]> {
+    const response = await fetch(`http://localhost:3001/api/gifs?tag=${tag}`);
+    const res: GifDTO[] = await response.json();
+    return res;
+  }
+}
