+++ guifaffinity-front/src/infrastructures/GifDTO.ts
@@ -1,4 +1,4 @@
-export interface GifType {
+export interface GifDTO {
   id: string;
   src: string;
   likes: number;
