--- guifaffinity-front/src/repositories/FetchGifsRepository.ts
@@ -1,11 +1,0 @@
-import { GifType } from "../interfaces/gif.interface";
-import { GifsRepository } from "./GifsRepository";
-
-export class FetchGifsRepository implements GifsRepository {
-  async getGifs(tag: string): Promise<GifType[]> {
-    console.log("Si no Onintza no se ubica");
-    const response = await fetch(`http://localhost:3001/api/gifs?tag=${tag}`);
-    const res = await response.json();
-    return res;
-  }
-}
