--- guifaffinity-front/src/repositories/GifsRepository.ts
@@ -1,5 +1,0 @@
-import { GifType } from "../interfaces/gif.interface";
-
-export interface GifsRepository {
-  getGifs: (tag: string) => Promise<GifType[]>;
-}
