+++ guifaffinity-front/src/services/_di/GifsService.container.ts
@@ -1,0 +1,4 @@
+import { FetchGifsRepository } from "../../infrastructures/FetchGifsRepository";
+import { GifsService } from "../GifsService";
+
+export const gifsServiceContainer = new GifsService(new FetchGifsRepository());
